+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.28`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.29`
 - `gradle -> 7.5`
 - `org.jetbrains.kotlin:kotlin-gradle-plugin -> 1.7.10`