+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 
 ### Updated
 
+- `com.javiersc.semver:semver-core -> 0.1.0-beta.13`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.46`
 - `gradle -> 7.6`